# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for URL shortening with support for custom slugs",
        "title": "Zipway URL Shortener API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/resolve/{slug}": {
            "get": {
                "description": "Returns the target URL for a given slug. Public endpoint, no authentication required.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "links"
                ],
                "summary": "Resolve a shortened link",
                "parameters": [
                    {
                        "type": "string",
                        "example": "abc123",
                        "description": "Shortened link slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Target URL",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Link not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/shorten": {
            "post": {
                "description": "Create a new shortened link from a URL. Requires authentication. Optionally allows defining a custom slug. Reserved slugs (api, swagger, shorten, admin, health, metrics, docs, static, assets, favicon.ico) cannot be used.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "links"
                ],
                "summary": "Create a shortened link",
                "parameters": [
                    {
                        "description": "Link data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateShortLinkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Link created successfully",
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateShortLinkResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error or reserved slug",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Custom slug already exists",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/{slug}": {
            "get": {
                "description": "Redirects to the original URL associated with the provided slug",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "links"
                ],
                "summary": "Redirect to original URL",
                "parameters": [
                    {
                        "type": "string",
                        "example": "abc123",
                        "description": "Shortened link slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "301": {
                        "description": "Permanent redirect",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Link not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Link": {
            "type": "object",
            "properties": {
                "clicks": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "short_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.LinkStatus"
                },
                "target_url": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "domain.LinkStatus": {
            "type": "string",
            "enum": [
                "ACTIVE",
                "PAUSED"
            ],
            "x-enum-varnames": [
                "StatusActive",
                "StatusPaused"
            ]
        },
        "handlers.CreateShortLinkRequest": {
            "type": "object",
            "required": [
                "target_url"
            ],
            "properties": {
                "custom_slug": {
                    "type": "string",
                    "example": "my-custom-link"
                },
                "target_url": {
                    "type": "string",
                    "example": "https://example.com"
                }
            }
        },
        "handlers.CreateShortLinkResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "$ref": "#/definitions/domain.Link"
                },
                "short_url": {
                    "type": "string",
                    "example": "http://localhost:8080/abc123"
                }
            }
        },
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid input"
                }
            }
        }
    }
}